+++ bindings/vue/src/components/index.js
@@ -1,6 +1,7 @@
 export { default as VOnsNavigator } from './VOnsNavigator.vue';
 export { default as VOnsBackButton } from './VOnsBackButton.vue';
 export { default as VOnsTabbar } from './VOnsTabbar.vue';
+export { default as VOnsButton } from './VOnsButton.vue';
 export { default as VOnsRange } from './VOnsRange.vue';
 export { default as VOnsSwitch } from './VOnsSwitch.vue';
 export { default as VOnsPullHook } from './VOnsPullHook.vue';